--- packages/map/src/overlay/marker/bubble-marker/index.ts
@@ -1,2 +1,0 @@
-export { default as ScrapCircleMarker } from './scrap-circle-marker'
-export { default as SmallCircleMarker } from './small-circle-marker'
