+++ packages/map/src/overlay/markers/index.ts
@@ -1,2 +1,3 @@
 export * from './circle-marker'
 export * from './bubble-marker'
+export * from './overlay-base'
